Rewrote two types of pointers to structs in Olden/bh.
Replaced all uses of treeptr and bodyptr with mm_ptr<T>.
Currently bh can compile but failed llvm-lit test. It's hard to
analyze the reason as there are a moderate amount of changes
to the source code. The current plan is to rewrite a smaller
benchmark program and get back to this one later.