Finished rewriting Olden/bh.
In walksub.c, there was an assignment from a raw C pointer to an mm_ptr.
After fixed this assignment and the corresponding definition in
hgstruct, bh can compile and run with -O0, and -O3. Before this commit
bh can compile and run with -O0 but failed to run with -O1 or higher.
For some unknown reason(s), the memory address of on movaps instruction
in gravsub() was not aligned on a 16-byte boundary.