Fixed the checked version of bisort.
We had a really weird bug: before this commit the checked version of
the bisort program stopped compiling because the types of its declared
"malloc()" and defined "random()" functions conflict with these two
functions defined in libc. This makes sense. However, how come we
compiled and ran it without any problem before?